+++ src/tools.h
@@ -1,6 +1,6 @@
-int** allocTable(size_t n, size_t m);
-void freeMemory(size_t n, int** ptr);
-void shiftArray(size_t n, int* array);
+int** allocTable(int n, int m);
+void freeMemory(int n, int** ptr);
+void shiftArray(int n, int* array);
 
 int viagem(int** travel, clubes_t* clubes, int da_casa, int nemesis);
